requirements.txt
affine==2.3.0
asgiref==3.4.1
attrs==21.2.0
black==21.11b0
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Django==3.2.9
django-adminlte-3==0.1.6
django-environ==0.8.1
django-filter==21.1
django-leaflet==0.28.2
django-leaflet-admin-list==0.0.4
django-suit==2.0a1
djangorestframework==3.12.4
djangorestframework-gis==0.17
Fiona==1.8.20
flake8==4.0.1
fonttools==4.28.2
# GDAL==3.3.0
GeoAlchemy2==0.9.4
geopandas==0.10.2
geoserver-rest==2.1.4
greenlet==1.1.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.0
mccabe==0.6.1
munch==2.5.0
mypy-extensions==0.4.3
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pycurl==7.44.1
pyflakes==2.4.0
Pygments==2.10.0
pyparsing==3.0.6
pyproj==3.3.0
python-dateutil==2.8.2
pytz==2021.3
rasterio==1.2.10
regex==2021.11.10
requests==2.26.0
Rtree==0.9.7
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
Shapely==1.8.0
six==1.16.0
snuggs==1.4.7
SQLAlchemy==1.4.27
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
django-extra-fields==3.0.2
